Calculate Contiguous Cartogram Boundaries

This function has been renamed: Please use cartogram_cont() instead of cartogram().

cartogram(shp, ...)

Arguments

shp
SpatialPolygonDataFrame or an sf object

...
Arguments passed on to cartogram_cont

weight

Name of the weighting variable in x

itermax

Maximum iterations for the cartogram transformation, if maxSizeError ist not reached

maxSizeError

Stop if meanSizeError is smaller than maxSizeError

prepare

Weighting values are adjusted to reach convergence much earlier. Possible methods are "adjust", adjust values to restrict the mass vector to the quantiles defined by threshold and 1-threshold (default), "remove", remove features with values lower than quantile at threshold, "none", don't adjust weighting values

threshold

Define threshold for data preparation