Calculate Non-Contiguous Cartogram Boundaries

This function has been renamed: Please use cartogram_ncont() instead of nc_cartogram().

nc_cartogram(shp, ...)

Arguments

shp
SpatialPolygonDataFrame or an sf object

...
Arguments passed on to cartogram_ncont

weight

Name of the weighting variable in x

k

Factor expansion for the unit with the greater value

inplace

If TRUE, each polygon is modified in its original place, if FALSE multi-polygons are centered on their initial centroid