Calculate Non-Contiguous Cartogram Boundaries

This function has been renamed: Please use cartogram_ncont() instead of nc_cartogram().

Usage

nc_cartogram(shp, ...)

Arguments

shp

SpatialPolygonDataFrame or an sf object

...

Arguments passed on to cartogram_ncont

weight

Name of the weighting variable in x

k

Factor expansion for the unit with the greater value

inplace

If TRUE, each polygon is modified in its original place, if FALSE multi-polygons are centered on their initial centroid

n_cpu

Number of cores to use. Defaults to "respect_future_plan". Available options are: * "respect_future_plan" - By default, the function will run on a single core, unless the user specifies the number of cores using plan (e.g. `future::plan(future::multisession, workers = 4)`) before running the `cartogram_ncont` function. * "auto" - Use all except available cores (identified with availableCores) except 1, to keep the system responsive. * a `numeric` value - Use the specified number of cores. In this case `cartogram_ncont` will use set the specified number of cores internally with `future::plan(future::multisession, workers = n_cpu)` and revert that back by switching the plan back to whichever plan might have been set before by the user. If only 1 core is set, the function will not require `future` and `future.apply` and will run on a single core.

show_progress

A `logical` value. If TRUE, show progress bar. Defaults to TRUE.